fix: eliminar 73 argument.type y 249 errores en cascada (604 errores PHPStan, -322)
Fixes de tipos en RemoteSensing, controllers, jobs y commands:
- (float) casts en ndvi_mean, ndwi, temperature, soil_moisture (string|null → float)
- (int) casts en ceil/floor → substr, daysRemaining (float → int)
- /** @var Plot */ PHPDoc donde relaciones devolvían Eloquent\Model genérico
- Carbon::parse() en beta_ends_at (string|null → Carbon|null)
- $user->id en lugar de $user donde se esperaba int
- (string) cast en header() con int
- whereRaw() reemplazando whereColumn(DB::raw()) (Expression → string)

Los PHPDoc de Plot arrastraron -249 property.notFound adicionales.

926 → 604 errores (-322).

+++ app/Services/RemoteSensing/Calculators/IrrigationCalculator.php
@@ -43,7 +43,7 @@ class IrrigationCalculator
         $irrigationNeed = max(0, $weeklyNeed - $effectivePrecip - $soilReserve);
 
         // Convert to liters per hectare (1mm = 10,000 L/ha)
-        $litersPerHa = round($irrigationNeed * 10000);
+        $litersPerHa = (int) round($irrigationNeed * 10000);
 
         // Generate recommendation
         [$text, $color, $bg] = $this->getRecommendation($irrigationNeed);
